+++ src/test/java/dinkplugin/notifiers/PetNotifierTest.java
@@ -45,6 +45,9 @@ class PetNotifierTest extends MockedNotifierTest {
 
     @Test
     void testNotify() {
+        // update mocks
+        when(config.petNotifyMessage()).thenReturn("%USERNAME% %GAME_MESSAGE%");
+
         // send fake message
         notifier.onChatMessage("You feel something weird sneaking into your backpack.");
         IntStream.rangeClosed(0, MAX_TICKS_WAIT).forEach(i -> notifier.onTick());
@@ -55,7 +58,7 @@ class PetNotifierTest extends MockedNotifierTest {
             false,
             NotificationBody.builder()
                 .extra(new PetNotificationData(null, null, false))
-                .text(buildTemplate(PLAYER_NAME + " got a pet"))
+                .text(buildTemplate(PLAYER_NAME + " feels something weird sneaking into their backpack"))
                 .type(NotificationType.PET)
                 .build()
         );
@@ -63,6 +66,9 @@ class PetNotifierTest extends MockedNotifierTest {
 
     @Test
     void testNotifyDuplicate() {
+        // update mocks
+        when(config.petNotifyMessage()).thenReturn("%USERNAME% %GAME_MESSAGE%");
+
         // send fake message
         notifier.onChatMessage("You have a funny feeling like you would have been followed...");
         IntStream.rangeClosed(0, MAX_TICKS_WAIT).forEach(i -> notifier.onTick());
@@ -73,7 +79,7 @@ class PetNotifierTest extends MockedNotifierTest {
             false,
             NotificationBody.builder()
                 .extra(new PetNotificationData(null, null, true))
-                .text(buildTemplate(PLAYER_NAME + " got a pet"))
+                .text(buildTemplate(PLAYER_NAME + " has a funny feeling like they would have been followed..."))
                 .type(NotificationType.PET)
                 .build()
         );
@@ -259,11 +265,14 @@ class PetNotifierTest extends MockedNotifierTest {
 
     @Test
     void testNotifyOverride() {
+        // update mocks
+        when(config.petNotifyMessage()).thenReturn("%USERNAME% %GAME_MESSAGE%");
+
         // define url override
         when(config.petWebhook()).thenReturn("https://example.com");
 
         // send fake message
-        notifier.onChatMessage("You feel something weird sneaking into your backpack.");
+        notifier.onChatMessage("You have a funny feeling like you're being followed.");
         IntStream.rangeClosed(0, MAX_TICKS_WAIT).forEach(i -> notifier.onTick());
 
         // verify handled at override url
@@ -272,7 +281,7 @@ class PetNotifierTest extends MockedNotifierTest {
             false,
             NotificationBody.builder()
                 .extra(new PetNotificationData(null, null, false))
-                .text(buildTemplate(PLAYER_NAME + " got a pet"))
+                .text(buildTemplate(PLAYER_NAME + " has a funny feeling like they're being followed"))
                 .type(NotificationType.PET)
                 .build()
         );