+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Unreleased
 
-- Minor: Indicate whether pets are already owned in pet notifications. (#303)
+- Minor: Send kill count notifications for Penance Queen kills. (#304)
+- Minor: Include popup notification widget in collection log screenshots. (#309)
+- Minor: Support new thread creation for Discord forum channels. (#302)
+- Minor: Indicate whether pets are already owned in pet notifications. (#303, #307)
 - Minor: Include NPC kill count in loot notifications. (#299)
 - Minor: Add option to disable price in Group Storage notifications. (#298)
 - Minor: Add `%TOTAL_LEVEL%` message template for level notifications. (#300)